Upgrade Guice and RoboGuice.
There is no large change in the production code. Custom test runner for each
tests is no longer needed. Instead each test should extend RoboGuiceTest and
should provide AbstractModule and run with HQTPTestRunner.

Since we change the structure of test, the injection of test class instance
and activity class instance are done in the activity's context scope. This
timing of injection is conform with the timing of injection when you run the
application in the real device.