Change implementation of TimelineRecurringUpdaterImpl.
TimelineRecurringUpdater need to be ready for restart. This is requried
because activities occasionally pause. This forces to change the way to
implement updater. Just canceling the task is not enough.

Moreover, executor's cancellation is difficult in generalto use because it is
difficult to find the right way to cancel on-going processing and to clean up
everything at the arbitrary time. Because of this, this commit lessen the
direct use of executor framework.